# mtest: fix test output issues (in console)