I was annoyed by e.g. the meson's test subcommand's output's usage of the scissors emoji that never displayed properly in my development environments and set out to improve the look and feel of the test subcommand's output. 

In the commit message I write:

> This change set aims to fix various "issues" seen with the current
> implementation. The changes can be summarized with the following list:
> 
> * Replace emojis and spinners with multiline status displaying the name
>   and running time of each currently running test.
> * The test output (especially in verbose mode or when multiple failing
>   tests' output gets printed out) can get confusing. Try to make the
>   output easier to read and grasp. Most notable change here is the
>   addition of the test number to the beginning of each printed line.
> * Print exit details (i.e. exit code) of the test in verbose mode.
> * Try to make the verbose "live" output from tests to match the look and
>   feel of otherwise produced (verbose) test output.

which quite well describes the changes I believe (hope). A "picture" is however worth a thousand words:

  https://asciinema.org/a/dGOULiN7z9t0OJXrAno570W7T

In my opinion the end result is better that the original. This is however quite a subjective matter and you might disagree...

In the case that people agree with me on the above I still feel that it would be good if other people tried this out as well before merging. AFAIK the "test output" is not really tested via meson's own tests and as I can check only so many corner cases on my own there are bound to be issues that got overlooked. Especially if someone has strong opinions on how the subtests should be reported I'd love to hear back from you.