@jpakkane 2021-09-07
Looks nice. Has this been tested on Windows? Its terminal is wonky.