@trhd 2021-09-07
> Has this been tested on Windows? Its terminal is wonky.
Not at all. Unfortunately I dont have access to a Windows machine (well I suppose I could set up a virtual machine of some sort) so that is definitely something I could use some help on testing.