@anarazel 2021-09-16
I see two issues trying this out:

1) summary() referencing to dependencies are broken:
```
  External Libraries
    zlib           : [<mesonbuild.mlog.AnsiDecorator object at 0x7f68a8ed7730>, ' ', <mesonbuild.mlog.AnsiDecorator object at 0x7f68a8a2ad00>]
```
on master that is
```
  External Libraries
    zlib           : YES 1.2.11
```

2) The new output doesn't work for larger concurrency. On a 40 core workstation, in a project with 98 tests, it printed out 1577 lines. With status updates repeating a lot of information:

```
13/98 postgresql:tap+pg_test_timing / pg_test_timing/t/001_basic.pl                           OK                0.11s
      postgresql:tap+pg_verifybackup / pg_verifybackup/t/001_basic.pl                                           0/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/008_min_recovery_point.pl                                          0/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/007_standby_source.pl                                              0/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/005_same_timeline.pl                                               1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/004_pg_xlog_symlink.pl                                             1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/003_extrafiles.pl                                                  1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/002_databases.pl                                                   1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/001_basic.pl                                                       1/ 300s
      postgresql:tap+pg_basebackup / pg_basebackup/t/010_dump_connstr.pl                                        1/ 300s
      postgresql:tap+pg_basebackup / pg_basebackup/t/003_pg_dump_with_server.pl                                 1/ 300s
      postgresql:tap+pg_basebackup / pg_basebackup/t/002_pg_dump.pl                                             2/ 300s
      postgresql:tap+pg_verifybackup / pg_verifybackup/t/002_algorithm.pl                                       0/ 300s
      postgresql:tap+pg_verifybackup / pg_verifybackup/t/001_basic.pl                                           0/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/008_min_recovery_point.pl                                          0/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/007_standby_source.pl                                              0/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/005_same_timeline.pl                                               1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/004_pg_xlog_symlink.pl                                             1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/003_extrafiles.pl                                                  1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/002_databases.pl                                                   1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/001_basic.pl                                                       1/ 300s
      postgresql:tap+pg_basebackup / pg_basebackup/t/010_dump_connstr.pl                                        1/ 300s
      postgresql:tap+pg_basebackup / pg_basebackup/t/003_pg_dump_with_server.pl                                 2/ 300s
      postgresql:tap+pg_basebackup / pg_basebackup/t/002_pg_dump.pl                                             2/ 300s
14/98 postgresql:tap+pg_verifybackup / pg_verifybackup/t/001_basic.pl                         OK                0.14s
      postgresql:tap+pg_verifybackup / pg_verifybackup/t/002_algorithm.pl                                       0/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/008_min_recovery_point.pl                                          0/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/007_standby_source.pl                                              0/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/005_same_timeline.pl                                               1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/004_pg_xlog_symlink.pl                                             1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/003_extrafiles.pl                                                  1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/002_databases.pl                                                   1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/001_basic.pl                                                       2/ 300s
      postgresql:tap+pg_basebackup / pg_basebackup/t/010_dump_connstr.pl                                        2/ 300s
      postgresql:tap+pg_basebackup / pg_basebackup/t/003_pg_dump_with_server.pl                                 2/ 300s
      postgresql:tap+pg_basebackup / pg_basebackup/t/002_pg_dump.pl                                             2/ 300s
      postgresql:tap+pg_verifybackup / pg_verifybackup/t/003_corruption.pl                                      0/ 300s
      postgresql:tap+pg_verifybackup / pg_verifybackup/t/002_algorithm.pl                                       0/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/008_min_recovery_point.pl                                          0/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/007_standby_source.pl                                              0/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/005_same_timeline.pl                                               1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/004_pg_xlog_symlink.pl                                             1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/003_extrafiles.pl                                                  1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/002_databases.pl                                                   1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/001_basic.pl                                                       2/ 300s
      postgresql:tap+pg_basebackup / pg_basebackup/t/010_dump_connstr.pl                                        2/ 300s
      postgresql:tap+pg_basebackup / pg_basebackup/t/003_pg_dump_with_server.pl                                 2/ 300s
      postgresql:tap+pg_basebackup / pg_basebackup/t/002_pg_dump.pl                                             2/ 300s
      postgresql:tap+pg_verifybackup / pg_verifybackup/t/003_corruption.pl                                      0/ 300s
      postgresql:tap+pg_verifybackup / pg_verifybackup/t/002_algorithm.pl                                       0/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/008_min_recovery_point.pl                                          0/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/007_standby_source.pl                                              0/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/005_same_timeline.pl                                               1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/004_pg_xlog_symlink.pl                                             1/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/003_extrafiles.pl                                                  2/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/002_databases.pl                                                   2/ 300s
      postgresql:tap+pg_rewind / pg_rewind/t/001_basic.pl                                                       2/ 300s
      postgresql:tap+pg_basebackup / pg_basebackup/t/010_dump_connstr.pl                                        2/ 300s
      postgresql:tap+pg_basebackup / pg_basebackup/t/003_pg_dump_with_server.pl                                 2/ 300s
      postgresql:tap+pg_basebackup / pg_basebackup/t/002_pg_dump.pl                                             2/ 300s
```
whereas on master it ends up with roughly one line per finished test, and the summary.

3) This branch lost the information about subtest information, as e.g. present for tap:
master:
```
 2/98 postgresql:tap+pg_checksums / pg_checksums/t/001_basic.pl                               OK                0.16s   8 subtests passed
```
PR:
```
 2/98 postgresql:tap+pg_checksums / pg_checksums/t/001_basic.pl                               OK                0.17s
```